+++ pagamento-service/src/main/java/br/com/fiap/payment/infra/gateway/http/dto/ProcPagHttpRequest.java
@@ -1,11 +1,9 @@
 package br.com.fiap.payment.infra.gateway.http.dto;
 
-import java.math.BigDecimal;
-
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record ProcPagHttpRequest(
     @JsonProperty("pagamento_id") String pagamentoId,
     @JsonProperty("cliente_id") String clienteId,
-    @JsonProperty("valor") BigDecimal valor
+    @JsonProperty("valor") Long valor
 ) {}
